+++ src/mailman/docs/NEWS.rst
@@ -109,6 +109,7 @@ Bugs
  * Fix ``mailman stop`` not stopping some runners due to PEP 475 interaction.
    (Closes: #255)
  * Update documentation links for ``config.cfg`` settings.  (Closes: #306)
+ * Disallow problematic characters in listnames.  (Closes: #311)
 
 Configuration
 -------------
@@ -119,6 +120,8 @@ Configuration
    rules is not yet exposed through the REST API.  Given by Aurélien Bompard.
  * The default languages from Mailman 2.1 have been ported over.  Given by
    Aurélien Bompard.
+ * There is now a configuration setting to limit the characters that can be
+   used in list names.
 
 Command line
 ------------